+++ tests/dataset_test.py
@@ -43,7 +43,7 @@ def test_add_local_file_to_new_dataset(typ, fs):
     assert f.remote_access_path(dset.source_folder) is None
     assert f.local_path == Path("local/folder/data.dat")
     assert f.size == file_data["size"]
-    assert f.checksum_algorithm == "md5"
+    assert f.checksum_algorithm == "blake2b"
 
     assert abs(file_data["creation_time"] - f.creation_time) < timedelta(seconds=1)
     assert abs(file_data["creation_time"] - f.make_model().time) < timedelta(seconds=1)
@@ -72,14 +72,14 @@ def test_add_multiple_local_files_to_new_dataset(typ, fs):
     assert f0.remote_access_path(dset.source_folder) is None
     assert f0.local_path == Path("common/location1/data.dat")
     assert f0.size == file_data0["size"]
-    assert f0.checksum_algorithm == "md5"
+    assert f0.checksum_algorithm == "blake2b"
 
     assert not f1.is_on_remote
     assert f1.is_on_local
     assert f1.remote_access_path(dset.source_folder) is None
     assert f1.local_path == Path("common/song.mp3")
     assert f1.size == file_data1["size"]
-    assert f1.checksum_algorithm == "md5"
+    assert f1.checksum_algorithm == "blake2b"
 
 
 @pytest.mark.parametrize("typ", (DatasetType.RAW, DatasetType.DERIVED))
@@ -107,14 +107,14 @@ def test_add_multiple_local_files_to_new_dataset_with_base_path(typ, fs):
     assert f0.remote_access_path(dset.source_folder) is None
     assert f0.local_path == Path("common/location1/data.dat")
     assert f0.size == file_data0["size"]
-    assert f0.checksum_algorithm == "md5"
+    assert f0.checksum_algorithm == "blake2b"
 
     assert not f1.is_on_remote
     assert f1.is_on_local
     assert f1.remote_access_path(dset.source_folder) is None
     assert f1.local_path == Path("common/song.mp3")
     assert f1.size == file_data1["size"]
-    assert f1.checksum_algorithm == "md5"
+    assert f1.checksum_algorithm == "blake2b"
 
 
 @pytest.mark.parametrize("typ", (DatasetType.RAW, DatasetType.DERIVED))